Main file created
Just for a heads up the small data set I made and the predictions this algorithm will make are based off of AMERICAN ADULT averages!!! In this python file I first import the machine learning library scikit-learn and the scientific computing library numpy as my dependencies . Then I parse the feature data set text file into a 2 dimensional array of integers. Then I do the same for the labels data set except I keep it as a string data type. After that I specify the use of the decision tree classifier algorithm that will be used to determine whether the future inputs describe a male or a female. Then I fit my variables onto a graph for the algorithm to use. Lastly, the user is prompted to enter in their qualities to classify their gender.